pyproject.toml
[tool.poetry]
name = "tutorial"
version = "0.1.0"
description = "A simple project from django-ninja tutorial"
authors = ["jlucasoliveira <lucasosilva68@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
django-ninja = "^0.17.0"
djangorestframework = "^3.13.1"
orjson = "^3.6.7"
PyJWT = {extras = ["crypto"], version = "^2.3.0"}
python-decouple = "^3.6"
debugpy = "^1.6.0"
pydantic = "^1.9.0"
serpy = "^0.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
